# frontend/src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Settly Core API",
    "version": "1.0.0",
    "description": "Authoritative HTTP API contract for the Settly real estate platform. Implements RFC 9457 error taxonomy, strict validation, and bilingual domain entities."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "settly_session",
        "description": "Better Auth HTTP-only secure session cookie"
      }
    },
    "schemas": {
      "BaseProblem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Relative URI identifier of the specific problem type (machine-readable code)",
            "example": "/errors/validation-failed"
          },
          "title": {
            "type": "string",
            "description": "Short human-readable developer summary of the problem type",
            "example": "Validation Failed"
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code generated by the origin server",
            "example": 422
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence of the problem",
            "example": "One or more request parameters failed validation schema checks."
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence of the problem",
            "example": "/api/v1/areas?level=INVALID"
          },
          "requestId": {
            "type": "string",
            "description": "Authoritative request correlation ID (UUIDv7) for telemetry tracing",
            "example": "0191eb45-8f67-73d8-9db8-bc234a9e51c8"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "Structured machine-readable parameters for localized client error rendering"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "requestId"
        ]
      },
      "ValidationProblem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProblem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "/errors/validation-failed"
                ]
              },
              "status": {
                "type": "number",
                "enum": [
                  422
                ]
              },
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string",
                      "example": "level"
                    },
                    "code": {
                      "type": "string",
                      "example": "invalid_enum_value"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid enum value"
                    },
                    "params": {
                      "type": "object",
                      "additionalProperties": {
                        "nullable": true
                      }
                    }
                  },
                  "required": [
                    "path",
                    "code"
                  ]
                },
                "description": "Array of validation issues per field"
              }
            },
            "required": [
              "errors"
            ]
          }
        ]
      },
      "NotFoundProblem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProblem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "/errors/not-found"
                ]
              },
              "status": {
                "type": "number",
                "enum": [
                  404
                ]
              }
            }
          }
        ]
      },
...
      "UnauthenticatedProblem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseProblem"
          },
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "/errors/unauthenticated"
                ]
              },
              "status": {
                "type": "number",
                "enum": [
                  401
                ]
              }
            }
          }
        ]
      },
...
      "UserDeviceRegistrationInput": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 10,
            "maxLength": 500,
            "description": "FCM (Firebase Cloud Messaging) device or browser push token",
            "example": "fcm_token_sample_abc1234567890"
          },
          "platform": {
            "type": "string",
            "enum": [
              "IOS",
              "ANDROID",
              "WEB"
            ],
            "description": "Target platform for push notification dispatch",
            "example": "WEB"
          }
        },
        "required": [
          "token",
          "platform"
        ]
      },
      "UserDeviceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "token",
          "platform",
          "createdAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/me/devices": {
      "post": {
        "summary": "Register or update FCM device token for push notifications",
        "description": "Associates a Web Push or Mobile FCM token with the authenticated user.",
        "tags": [
          "Identity",
          "Notifications"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDeviceRegistrationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDeviceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedProblem"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/devices/{token}": {
      "delete": {
        "summary": "Deregister FCM device token",
        "description": "Removes push notification token upon user logout or notification revocation.",
        "tags": [
          "Identity",
          "Notifications"
        ],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "token",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Device deregistered successfully (no content)"
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedProblem"
                }
              }
            }
          },
          "404": {
            "description": "Device token not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundProblem"
                }
              }
            }
          }
        }
      }
    }
  }
}